refactor: Generalize rendering recreation
The previous `handleRecreateInitialRendering` function was specific to recreating the 'Front Exterior' rendering. This commit generalizes the function to `handleRecreateRendering` which accepts a `category` parameter. This allows for recreation of other rendering types, such as 'Back Exterior', improving the flexibility of the application. The `ResultsPage` component has also been updated to pass the rendering category to the new function.